+++ src/com/ansorgit/plugins/bash/editor/inspections/inspections/SimpleArrayUseInspection.java
@@ -17,6 +17,7 @@ package com.ansorgit.plugins.bash.editor.inspections.inspections;
 
 import com.ansorgit.plugins.bash.lang.psi.BashVisitor;
 import com.ansorgit.plugins.bash.lang.psi.api.BashString;
+import com.ansorgit.plugins.bash.lang.psi.api.arithmetic.ArithmeticExpression;
 import com.ansorgit.plugins.bash.lang.psi.api.vars.BashVar;
 import com.ansorgit.plugins.bash.lang.psi.api.vars.BashVarDef;
 import com.ansorgit.plugins.bash.lang.psi.util.BashPsiUtils;
@@ -25,6 +26,7 @@ import com.intellij.codeInspection.ProblemHighlightType;
 import com.intellij.codeInspection.ProblemsHolder;
 import com.intellij.psi.PsiElementVisitor;
 import org.jetbrains.annotations.NotNull;
+import sun.tools.tree.BinaryArithmeticExpression;
 
 /**
  * This inspection detects use of array variables without array element qualifiers.
@@ -42,7 +44,10 @@ public class SimpleArrayUseInspection extends LocalInspectionTool {
                         if (!definition.isArray()) {
                             holder.registerProblem(var, "Array use of non-array variable", ProblemHighlightType.WEAK_WARNING);
                         }
-                    } else if (definition.isArray() && !BashPsiUtils.hasParentOfType(var, BashString.class, 5)) {
+                    } else if (definition.isArray()
+                            && !BashPsiUtils.hasParentOfType(var, BashString.class, 5)
+                            && !BashPsiUtils.hasParentOfType(var, ArithmeticExpression.class, 5)) {
+
                         holder.registerProblem(var, "Simple use of array variable", ProblemHighlightType.WEAK_WARNING);
                     }
                 }